requirements.txt
aws-cdk-lib==2.39.1
constructs>=10.0.0,<11.0.0
